requirements.txt
streamlit
google-generativeai
python-dotenv
ipython
PyPDF2
chromadb
